format and move func outside of pymodule
