+++ jsql/myenv/pyvenv.cfg
@@ -1,0 +1,5 @@
+home = /opt/homebrew/opt/python@3.11/bin
+include-system-site-packages = false
+version = 3.11.5
+executable = /opt/homebrew/Cellar/python@3.11/3.11.5/Frameworks/Python.framework/Versions/3.11/bin/python3.11
+command = /opt/homebrew/opt/python@3.11/bin/python3.11 -m venv /Users/ddak/RustroverProjects/jsql/jsql/myenv
